// types/BusinessFormData.ts
// types/BusinessFormData.ts
export type BusinessFormData = {
  business_name: string;
  business_type: string;
  category: string;
  phone_number: string;
  business_email: string;
  barangay: string;
  city: string;
  province: string;
  postal_code: string;
  description: string;
  instagram_url: string;
  twitter_url: string;
  facebook_url: string;
  latitude: string;
  longitude: string;
  price_range: string;
};